requirements.txt
# Core ML Framework
autogluon.timeseries==1.1.0

# MLOps Tools
mlflow==2.9.2
metaflow==2.11.7

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Visualization (optional, for MLflow artifacts)
matplotlib==3.8.2
seaborn==0.13.0

# Development
jupyter==1.0.0